tidy layout (#7)
* Start tidying rendering.

- Remove renderPartial(). Use render() for everything _except_
  where we have to re-render the nav ie. renderWithNav().

* Testing monolithic loading.

Example of monolithic loading of a hypothetical settings page, with a
tab control. The handler needs to know about the models it will show in
it's template, and it turns into a bit of a mess. The conditional in the
template is really bad.

It's best to go back to loading things from the template, instead of
trying to populate everything and show in one roundtrip. What it costs
us in network, we gain back in responsiveness and separation.