stuff (#8)
* Add visibility check event filter.

* Some tidies.

* Clean up added extra block data to render.

Rather than creating special functions for different extra data to be
rendered with the template (eg. navigation bar or other oob stuff),
change render() to allow specifying the bits of data to include.

These are pulled in on rendering, which feels a little bit late, but
this keeps the endpoints that call render() clean. Will need to be
reviewed to see if there is a better pattern, but this should do for
now.

* Scroll at the right time.

* Fix dodgy settings page.